requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.3
dj-database-url==2.1.0
Django==5.0.5
django-cors-headers==4.3.1
django-timezone-field==6.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
kombu==5.3.7
MarkupSafe==2.1.5
numpy==1.26.4
openapi-codec==1.3.2
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
requests==2.31.0
setuptools==69.5.1
simplejson==3.19.2
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
